ABOUTME.md
Open this! you can get to know me a little and I have a given my IG username. POV don't need to follow me but you can request a message on it we can communicate. I'm a relationship advisor as a parttimer, and hopefully please try to help me to learn programming from noob to Pro. Thank you.